fix compiler warnings

+++ src/main.c
@@ -24,7 +24,7 @@ int main(int c, char **v)
     if (c == 2 && !strncmp(v[1], "-help", 5))
         show_help(0);
     p = strtol(v[1], NULL, 10);
-    s.res.port = (p > 0 && p < 65535) ? p : (int)errb("Invalid port number");
+    s.res.port = (p > 0 && p < 65535) ? p : (long)errb("Invalid port number");
     s.res.home = chdir(v[2]) != -1 ? v[2] : errb(strerror(errno));
     try_init_server(&s);
     run_server(&s);
